Add ContainerTile & an example inventory class for it
Cleaned up SimpleTile a bit, added some TODOs to consider later on, added a ContainerTile which should be able to store inventories and their contents. An example/fallback inventory class has also been introduced to show the users how to create their own inventories for the ContainerTile, and to keep a fallback value for the ContainerTile's inventory.